refactor(web): render settings into render_settings.rs; silence wasm-only dead code
- web lib.rs ratchets to 2151 (the settings block from the FFI-gap fix
  moves to its own module; fixes the ffi-parity CI failure on the
  previous commit, which tripped the line ratchet)
- encode_png_simple cfg'd to native (feeds the file-writing screenshot
  path; was a permanent warning on every wasm build)
