+++ tools/file-lines-baseline.json
@@ -1,6 +1,6 @@
 {
   "native/shared/src/renderer/mod.rs": 13602,
   "native/shared/src/renderer/shaders.rs": 5861,
-  "native/web/src/lib.rs": 2146,
+  "native/web/src/lib.rs": 2151,
   "tools/bloom-reference/src/main.rs": 2330
 }
